1.微信登陆bug 2.聊天语音气泡

+++ PeachTravel/IM/Chat/ChatView/MessageView/MessageModel.h
@@ -55,7 +55,7 @@
 //audio
 @property (nonatomic, strong) NSString *localPath;
 @property (nonatomic, strong) NSString *remotePath;
-@property (nonatomic) NSInteger time;
+@property (nonatomic) float time;
 @property (nonatomic) BOOL isPlaying;
 @property (nonatomic) BOOL isPlayed;
